implement environment variable mechanics for v2
App#env now returns a ChattyHash, which acts like a Hash but will report back
to App when things happen to it.

This is used to call App#env=, which for v1 saves it as the legacy variable
list form, and for v2, sets it as a Hash the diff.

Change-Id: I0c1c2e4f4e4da27194a2c2debc14733063d0de36