Use password grant, store all token data
Previously, we stored only the access token, which is a
base64-encoded string consisting of two JSON hashes. We
now store all of the data returned by the response from
the UAA, including the refresh token.

Also, we no longer lazily decode the access token. Instead,
the UAAClient decodes it as soon as it gets the response,
and adds the decoded data to the `token` hash.